patient-management/requirements.txt
flask 
flask_sqlalchemy
psycopg2-binary
Flask-Migrate
python-dotenv
